Moved FormPage file to its own Folder
