+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import Home from './Pages/Home/Home';
 import GradHatLoader from './components/Loader/GradHatLoader';
-import FormPage from './Pages/Home/Formpage';
+import FormPage from './Pages/FormPage/Formpage';
 import { Router, Route, Switch } from 'react-router';
 import { createBrowserHistory } from 'history';
